Move internship and user profile packages to a broader model package
