+++ src/main/java/seedu/internsprint/logic/command/ByeCommand.java
@@ -1,7 +1,7 @@
 package seedu.internsprint.logic.command;
 
-import seedu.internsprint.internship.InternshipList;
-import seedu.internsprint.userprofile.UserProfile;
+import seedu.internsprint.model.internship.InternshipList;
+import seedu.internsprint.model.userprofile.UserProfile;
 import seedu.internsprint.util.InternSprintLogger;
 
 import java.util.ArrayList;
