+++ src/main/java/seedu/internsprint/logic/command/internship/AddGeneralInternshipCommand.java
@@ -1,7 +1,7 @@
 package seedu.internsprint.logic.command.internship;
 
-import seedu.internsprint.internship.GeneralInternship;
-import seedu.internsprint.internship.Internship;
+import seedu.internsprint.model.internship.GeneralInternship;
+import seedu.internsprint.model.internship.Internship;
 
 import java.util.Set;
 import java.util.logging.Level;
