+++ src/main/java/seedu/internsprint/logic/command/internship/AddInternshipCommand.java
@@ -2,9 +2,9 @@ package seedu.internsprint.logic.command.internship;
 
 import seedu.internsprint.logic.command.Command;
 import seedu.internsprint.logic.command.CommandResult;
-import seedu.internsprint.internship.Internship;
-import seedu.internsprint.internship.InternshipList;
-import seedu.internsprint.userprofile.UserProfile;
+import seedu.internsprint.model.internship.Internship;
+import seedu.internsprint.model.internship.InternshipList;
+import seedu.internsprint.model.userprofile.UserProfile;
 import seedu.internsprint.util.InternSprintLogger;
 
 import java.util.ArrayList;
