+++ src/main/java/seedu/internsprint/model/internship/InternshipList.java
@@ -1,4 +1,4 @@
-package seedu.internsprint.internship;
+package seedu.internsprint.model.internship;
 
 import seedu.internsprint.storage.StorageHandler;
