+++ src/test/java/seedu/internsprint/logic/command/internship/DeleteCommandTest.java
@@ -3,11 +3,8 @@ package seedu.internsprint.logic.command.internship;
 import org.junit.jupiter.api.Test;
 
 import seedu.internsprint.logic.command.CommandResult;
-import seedu.internsprint.internship.InternshipList;
-import seedu.internsprint.logic.command.internship.AddGeneralInternshipCommand;
-import seedu.internsprint.logic.command.internship.AddHardwareInternshipCommand;
-import seedu.internsprint.logic.command.internship.DeleteCommand;
-import seedu.internsprint.userprofile.UserProfile;
+import seedu.internsprint.model.internship.InternshipList;
+import seedu.internsprint.model.userprofile.UserProfile;
 import seedu.internsprint.util.InternSprintExceptionMessages;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
