+++ src/test/java/seedu/internsprint/logic/command/internship/ListCommandTest.java
@@ -3,10 +3,9 @@ package seedu.internsprint.logic.command.internship;
 import org.junit.jupiter.api.Test;
 
 import seedu.internsprint.logic.command.CommandResult;
-import seedu.internsprint.internship.InternshipList;
-import seedu.internsprint.internship.SoftwareInternship;
-import seedu.internsprint.logic.command.internship.ListCommand;
-import seedu.internsprint.userprofile.UserProfile;
+import seedu.internsprint.model.internship.InternshipList;
+import seedu.internsprint.model.internship.SoftwareInternship;
+import seedu.internsprint.model.userprofile.UserProfile;
 
 import java.util.List;
